rename file to allow for new lambda file
